Добавлены доработки Дениса

+++ EMessenger.Client/MainWindow.xaml.cs
@@ -74,7 +74,6 @@ namespace EMessenger.Client
       }
 
     }
-
     /// <summary>
     /// Произвести добавление чата на событие щелком ЛКМ по кнопке "Добавить общий чат"
     /// </summary>
@@ -82,7 +81,19 @@ namespace EMessenger.Client
     /// <param name="e"></param>
     private void BtnAddGeneralChat_Click(object sender, RoutedEventArgs e)
     {
-      messenger.AddGeneralChat();
+      // Создаем экземпляр диалогового окна
+      EMessenger.Client.Views.AddGeneralChatDialog dialog = new EMessenger.Client.Views.AddGeneralChatDialog();
+
+      // Покажем диалоговое окно
+      if (dialog.ShowDialog() == true)
+      {
+        // Если пользователь ввел имя чата и нажал "ОК", 
+        // создаем новый общий чат
+        string chatName = dialog.ChatName;
+
+        // Вызываем метод добавления общего чата в мессенджере
+        messenger.AddGeneralChat(chatName);
+      }
     }
 
     /// <summary>
@@ -106,11 +117,28 @@ namespace EMessenger.Client
 
     private void BtnAddGroupChat_Click(object sender, RoutedEventArgs e)
     {
-      messenger.AddGroupChat();
+      // Создаем экземпляр диалогового окна
+      EMessenger.Client.Views.AddGroupChatDialog dialog = new EMessenger.Client.Views.AddGroupChatDialog();
+
+      // Покажем диалоговое окно
+      if (dialog.ShowDialog() == true)
+      {
+        // Если пользователь ввел имя чата и нажал "ОК", 
+        // создаем новый общий чат
+        string chatName = dialog.ChatName;
+
+        // Вызываем метод добавления общего чата в мессенджере
+        messenger.AddGroupChat(chatName);
+      }
     }
 
     private void BtnDelChat_Click(object sender, RoutedEventArgs e)
     {
+      if (messenger.SelectedChat==null)
+      {
+        MessageBox.Show("Не выбран чат для удаления.");
+        return;
+      }
       messenger.DeleteChat();
     }
